implemented pointer tool
