+++ src/Tools/DeleteGeomsTool.ts
@@ -1,113 +1,23 @@
 import {
-  BaseTool,
   GeomItem,
   IntersectionData,
-  VRViewport,
   XRController,
   XRPoseEvent,
-  Xfo,
   ZeaMouseEvent,
   ZeaPointerEvent,
 } from '@zeainc/zea-engine'
 import { AppData } from '../../types/types'
-import { ToolManager } from './ToolManager'
-import { line, lineMaterial } from '../helpers/line'
 import { CustomGeom } from './CreateTools/CustomGeom'
 import DeleteGeomChange from './CreateTools/Change/DeleteGeomChange'
 import { Change, UndoRedoManager } from '../UndoRedo'
+import { PointerTool } from './VRTools/PointerTool'
 
-class DeleteGeomsTool extends BaseTool {
-  private appData: AppData
-
-  private vrViewport: VRViewport
-  private prevCursor: string
-
-  private defaultTaycastDist = 0.0
-  public raycastDist = 20.0
-  private bindControllerId: number
-
+class DeleteGeomsTool extends PointerTool {
   private pointerButtonHeld = false
   private change: Change
-  private pointerGeomItems: GeomItem[] = []
 
-  constructor(appData: AppData, toolManager: ToolManager) {
-    super()
-    this.appData = appData
-
-    this.appData.renderer.getXRViewport().then((xrvp) => {
-      if (xrvp instanceof VRViewport) {
-        this.vrViewport = xrvp
-      }
-    })
-    this.toolManager = toolManager
-  }
-
-  /**
-   * Enables tools usage. This method is called by either the Viewport when a tool is removed, or the ToolManage if it is installed.
-   */
-  activateTool(): void {
-    super.activateTool()
-
-    if (this.appData && this.appData.renderer) {
-      this.prevCursor = this.appData.renderer.getGLCanvas().style.cursor
-      this.appData.renderer.getGLCanvas().style.cursor = 'crosshair'
-    }
-
-    const bindController = (controller: XRController) => {
-      // The tool might already be deactivated.
-      if (this.pointerGeomItems[controller.id]) return
-
-      this.defaultTaycastDist = controller.raycastDist
-      controller.raycastDist = this.raycastDist
-
-      const pointerGeomItem = new GeomItem('PointerRay', line, lineMaterial)
-      pointerGeomItem.setSelectable(false)
-      const pointerXfo = new Xfo()
-      pointerXfo.sc.set(1, 1, this.raycastDist)
-      pointerGeomItem.localXfoParam.value = pointerXfo
-      controller.tipItem.addChild(pointerGeomItem, false)
-      this.pointerGeomItems[controller.id] = pointerGeomItem
-    }
-
-    if (this.vrViewport) {
-      for (const controller of this.vrViewport.getControllers()) {
-        bindController(controller)
-      }
-      this.bindControllerId = this.vrViewport.on('controllerAdded', (event) => bindController(event.controller))
-    }
-  }
-
-  /**
-   * The deactivateTool method.
-   */
-  deactivateTool(): void {
-    super.deactivateTool()
-    if (this.appData && this.appData.renderer) {
-      this.appData.renderer.getGLCanvas().style.cursor = this.prevCursor
-    }
-    const unbindController = (controller: XRController) => {
-      if (!this.pointerGeomItems[controller.id]) return
-      controller.tipItem.removeChildByHandle(this.pointerGeomItems[controller.id])
-
-      controller.raycastDist = this.defaultTaycastDist
-      this.pointerGeomItems[controller.id] = null
-    }
-    if (this.vrViewport) {
-      for (const controller of this.vrViewport.getControllers()) {
-        unbindController(controller)
-      }
-      this.vrViewport.removeListenerById('controllerAdded', this.bindControllerId)
-    }
-  }
-
-  private setPointerLength(length: number, controller: XRController): void {
-    const pointerGeomItem = this.pointerGeomItems[controller.id]
-    if (pointerGeomItem) {
-      const pointerLocalXfo = pointerGeomItem.localXfoParam.value
-
-      pointerLocalXfo.sc.set(1, 1, length / controller.getTipXfo().sc.z)
-      pointerGeomItem.localXfoParam.value = pointerLocalXfo
-    }
+  constructor(appData: AppData) {
+    super(appData)
   }
 
   private handleGeometryIntersection(intersectionData: IntersectionData) {
